moving WhiteflagMesage into core mod
