+++ src/whiteflag/tests.rs
@@ -1,4 +1,4 @@
-use crate::WhiteflagMessage;
+use crate::whiteflag::wf_core::message::WhiteflagMessage;
 
 #[cfg(test)]
 #[test]
@@ -67,3 +67,29 @@ fn test_deserialize_auth_message() {}
 
 #[test]
 fn test_decode_auth_message() {}
+
+/* #[test]
+fn text_decode_hex_message() {
+    let message_encoded: String = "57463130a6a1f7da7067d41891592131a12a60c9053b4eb0aefe6263385da9f5b789421e1d7401009841882148a800000114c1e596006f04c050eca6420084".to_string();
+    let field_values = vec![
+        "WF",
+        "1",
+        "0",
+        "1",
+        "M",
+        "4",
+        "3efb4e0cfa83122b242634254c1920a769d615dfcc4c670bb53eb6f12843c3ae",
+        "80",
+        "2013-08-31T04:29:15Z",
+        "P00D00H00M",
+        "22",
+        "+30.79658",
+        "-037.82602",
+        "8765",
+        "3210",
+        "042",
+    ];
+    let message = WhiteflagMessage::compile_auth_message(field_values.clone()).unwrap();
+
+    assert_eq!(message_encoded, message.encode().toHexString()); //, "Encoding should be correct"
+} */
